+++ src/davotools/__init__.py
@@ -3,7 +3,7 @@
 
 # %%
 ## imports
-import os, sys, types
+import os, types
 
 # %%
 
@@ -12,8 +12,7 @@ import os, sys, types
 
 # %%
 ## davotools submodules
-from davotools import minor, io
-from davotools import image
+from davotools import image, io, minor
 
 # %%
 ## davotools submodule: view
@@ -27,17 +26,7 @@ view.pd = _view.view_pd
 
 # %%
 ## (internal) to set the module path
-def _set_path(
-        echo: bool = True
-) -> tuple[str, str]:
-    """
-    an internal function to set the module path
-    Args:
-        echo: bool = True # whether to show internal details
-    Returns:
-        module_path: str # the full path of the module
-        module_dir: str # the directory of module_path
-    """
+def _set_path(echo: bool = True) -> tuple[str, str]:
     module_path = os.path.abspath(__file__)
     module_dir = os.path.dirname(module_path)
     if echo:
@@ -45,24 +34,7 @@ def _set_path(
     return module_path, module_dir
 
 # %%
-## to load image_old as image
-def load_image_old() -> None:
-    """
-    a function to replace davotools.image with davotools.image_old
-    Args: None
-    Returns: None
-    """
-    import importlib.util as _ilu
-    import os as _os
-    _path = _os.path.join( _os.path.dirname(__file__), '_backup', 'image_old--20260428.py' )
-    _spec = _ilu.spec_from_file_location('davotools.image_old', _path)
-    _image_old = _ilu.module_from_spec(_spec)
-    _spec.loader.exec_module(_image_old)
-    sys.modules[__name__].image = _image_old
-
-# %%
 ## initialization
-module_path, module_dir = _set_path()
+module_path, module_dir = _set_path(echo=False)
 
 # %%
-