+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 -. Suggestions from Claude Code are reflected.
 -. Some minor elements (such as doctrings) are fixed.
 
+##
+
 # Old versions of 2026
 
 ### 20260225 v0.0.4
@@ -28,6 +30,8 @@
 ### 20260116 v0.0.2
 -. to align functionality with other packages.
 
+
+##
 
 # Old versions of 2025